requirements.txt
pytest
pytest-cov
pylint
black
autopep8
requests
quixstreams
python-dotenv